requirements.txt
asgiref==3.8.1
channels==4.1.0
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
pillow==10.3.0
postgres==4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8